requirements.txt
fastapi>=0.109.0
uvicorn>=0.24.0
pymongo>=4.6.0
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
bcrypt>=4.0.1
python-dotenv>=1.0.0